Add files via upload
I'm just adding the sln folder to this update.  The latest updates are I have user interaction with half of the site, still working on the rest.  I have added a few pictures and descriptions for each tab that I have created.  As for the register page, I have added more fields for the user to add more personal information, which is linked to my database. Lastly, I have made a link so the user can register with their Facebook or Google account. My next task over the next few days is to send email verification to the user personal email so they can verify it that way.